backend/requirements.txt
fastapi==0.111.*
uvicorn[standard]==0.29.*
sqlalchemy>=2.0
asyncpg  # postgres driver
pandas   # excel parsing
python-multipart  # form uploads
celery[redis]
python-dotenv
billiard>=4.2.1,<5.0
###########
python-jose[cryptography]~=3.3
passlib[bcrypt]~=1.7
pydantic[email]~=2.7
pydantic-settings~=2.2     # ← new
alembic~=1.13
